Better neovim sync (#30)
Tested locally. Testing in GH CI is a bit of a pain. We should probably
force push to a dedicated branch on the mirror repo when we push to some
branch here. This also probably needs to be a Python script. For now I
think this cleans things up.

From:
https://github.com/hudson-trading/slang-server/commit/0464e0c9f5d601076feb7d536db8c62a9a3684b9